Improvements on the integration with the VSCode
Showing the results using a provider TextDocumentContentProvider avoids a big number of unsaved documents open.

The `cypher` document filtering for the `Run Query` was moved clause when on the command palette, so the option is not shown when it isn't correct case. This is the reason why the extension now dependson `jakeboone02.cypher-query-language`.

This change aslo adds a show result summary feature/option. It's configurable in the settings and it's usueful for writing queries.